Changed source image map layout and added metadata
The original layout was due to the issues with getting the os and
os_ver from the profile. the new method (see pull #1) simplifies
this, and so I've reduced the complexity in the image map now.

I've also pulled some more metadata from the profile to be passed
into the packer build template, this is due to the aquilon hook
scripts no longer defaulting to values, and so I've had to implement
that as well.